Feature ETP-4830: Show real backend error message on inline toggle fail
Checks: npm, testid, tests, regen, xml, pw-mocked, pw-integration
Hooks-Verified: v2 11edf03c ef9f82218537 f8ebd789efe2
