+++ tools/app-shell/src/components/contract-ui/__tests__/DetailView.lineSidebarSaveDeleteFlow.vitest.jsx
@@ -146,7 +146,8 @@ vi.mock('@/lib/documentTotals', () => ({
   resolveTotalDiscountPct: () => 0,
 }));
 
-vi.mock('@/lib/backendErrors.js', () => ({
+vi.mock('@/lib/backendErrors.js', async (importOriginal) => ({
+  ...(await importOriginal()),
   translateBackendError: (m) => m,
 }));
