+++ tools/app-shell/src/components/contract-ui/__tests__/DetailView.render.vitest.jsx
@@ -183,7 +183,8 @@ vi.mock('@/lib/documentTotals', () => ({
   resolveTotalDiscountPct: () => 0,
 }));
 
-vi.mock('@/lib/backendErrors.js', () => ({
+vi.mock('@/lib/backendErrors.js', async (importOriginal) => ({
+  ...(await importOriginal()),
   translateBackendError: (m) => m,
 }));
